requirements.txt
streamlit
torch
transformers
sentence-transformers
Pillow
numpy
pandas
scikit-learn
nltk
regex